Refactoring Mapa e Card de Unidades
This was made in order to reduce the response time for the map page
load. The backend was taking about 5 seconds to start answering to the
frontend. The refactor splitted some responsibilities about data
management and now the data for the Card itself is loaded just when the
user click on the "Saiba Mais" link to open the Unidade card.